Adds detector that checks accessibility for clickable views.
In accordance with GAR 1.4, we check that custom views handle
accessibility events by making sure that if onTouchEvent is overridden,
performClick is as well. We also want to make sure that performClick
is called from onTouchEvent and that performClick calls super.performClick.

In the future, this check should be extended to look at views
that use OnTouchListener, finishing what we can do with static analysis
on GAR 1.4.

Change-Id: I42e37f30d9e4afb4fd8257b8356f8446b8dbcbb8
Signed-off-by: Shannon Chen <shannonchen@google.com>