Do not fail build if ndk.dir/sdk.dir is not set in local.properties
When importing a project without local.properties in Android studio,
studio creates the local.properties but with only sdk.dir,
which leads to failing builds when ndk is requested.

Build should not fail if local.properties does not contain
sdk.dir/ndk.dir and it has been set as an environment variable
or as a system property.

Change-Id: I617bc684e1213f50f945353160f652a8cabcd9e9