Change library cache.
Instead of a single task per aar that needs unzipping,
we now have a task in each project using the aar. Using a single
task in the root project would break config on demand.

In order to still only unzip the aar once, we use a cache to
do the unzip, and all the tasks still unzip in the same location
in the root project.

Change-Id: Id85581764747fd0efafb7cb88feaf2da4f453026