Add relativePackage variable to the templates
The relativePackage variable can be set to a prefix which should be used
for XML attributes that can specify a package prefix (relative to the manifest
package). This lets us put ".MainActivity" in the manifest and layout files
rather than inlining the fully qualified name in each location.

For now (and for compatibility) this variable does not have to be provided
by the system; if not set, it will use the full package name instead. However,
a template context can specify for example "" as the prefix when generating
activities in the default application package, or a subpackage prefix like "foo"
when generating activities into the application package + "foo", or some other
fully qualified package when generating activities outside the manifest
package.

This CL also updates the old manifest merger to extract package names; it was
doing that it one of its process methods (used from ADT), but not in another one
(used from Studio.)

Change-Id: I1621135895a7a61e2533036f77cbbaa18be5c9de