Tweak Wear and TV templates
Insert tools:deviceIds attributes into the layouts such that we can (in the future)
automatically pick a compatible device for these. This is necessary because for example
for wear you typically have separate layouts for round and square, but this is *not*
required via resource qualifiers, so the normal picking machinery doesn't work. Instead
there is a third layout view stub custom view which dynamically picks one. Therefore,
with a deviceId tools attribute we can record the intended compatible device(s) for
the layout, and use that in the layout renderer's display picking code.

(That's not implemented yet, but the sooner we have these attributes in, the more
actual projects in the field will have the right metadata in them to get it right).

I also added in missing tools:context attributes for some of the layouts, and reordered
the attributes according to Studio's formatting rules such that as soon as the user
edits there aren't a lot of diffs.

Change-Id: I1a8c713842201a88b5b71c5871a44af2d414d765