Fix "use gradle" warning from lint command line runner
The lint command line runner doesn't work properly for Gradle
projects; you really need to run lint as part of Gradle when
you are analyzing Gradle projects.

There was a check in the lint runner which emitted a warning for
it, but it had two problems:

1) it did not always detect Gradle projects, and
2) it continued to report issues, which for a broken project
   could easily result in hundreds of issues, thereby effectively
   hiding the vital warning that you need to run "gradlew lint"
   instead.

This CL now hides all other issues such that only the "you should
use gradle lint instead" warning is shown, and it fixes the missing
identification of Gradle projects.

Change-Id: I42f157debda4cfdac9eff31b2075138a2d1af409