[alloc-tracker] Minor edits to AllocationInfo#filter to allow autocompletion.
Breaks AllocationInfo#filter into a method that returns whether the
allocated class matches the filter text and a method that returns a
list of the stack frames that match the filter text.  This allows, for
example, Android Studio's AllocationsViewPanel to know which string to
add to the autocomplete options of its filter field.

Change-Id: I0ba4b1c82b38d59a4a320f6a5f959ee0e3cdcd94
Signed-off-by: Rebecca Chen <rechen@google.com>