Initial implementation of dominator computation.
The dominator tree is the backbone for calculating objects' retained size and 
shortest paths to roots. Before we embark on a full Lengauer-Tarjan 
implementation, we start with a simple O(N^2) algorithm, which is fast in 
practice for medium-sized graphs. We will use this algorithm for testing the 
more sophisticated implementation later.

Minor tweaks were made to the visitor implementation, so that we can express 
both composite size and topological sort. It's unclear whether we want to keep 
composite size in the long term, because it's not as precise as the retained 
size metric, and it's as hard as transitive closure (O(N^3) will not be 
practical).

Change-Id: I87c3478280d6d1afd134a45fb653117250afe245