Introduce BaseVariantOutputData.
As a preparation for the new multi-apk splits, this changeset
introduces the data structure for variant outputs. It is already
present at the IDE model, so this adds it to the internal model
used to create tasks.

For now a single output is always created and all task creation
code was updated to use the single output.

The IDE model builder is updated in a more generic way and uses
all available outputs.

Change-Id: I2b0c0babb95fa2a2cda31992e636595f6011001b