Allows exporting PreferenceActivities if it is deemed safe.
This turns off the PreferenceActivityDetector lint check in cases
where the target SDK version of the project is 19 or greater
and isValidFragment is overridden by the activity in question.
The security issue that PreferenceActivityDetector checks for
was patched in version 19 and so the lint check was being
a little too aggressive before this change.

Change-Id: I32cf749254aed5ba0321288ea662723b9c7862a2
Signed-off-by: Shannon Chen <shannonchen@google.com>