Hprof performance: switch to memory-mapped files.
Setting the stage for more optimizations to follow, here we switch from
stream-based I/O to (seekable) byte buffers. As an example, this will allow
for lazier, on-demand reconstruction of field values for unreachable objects,
without storing more than an offset in the Java heap. Memory-mapped buffers
are not part of the Java heap.

A couple of parsing bugs were fixed while getting the parser to read a 2.4Gb
hprof sample: we don't care about the signed-ness of IDs because we never
interpret them, but we care about lengths and offsets.

Benchmarks: shaves off 10%-20% of wall-clock time for non-trivial heap dumps
 (tested with 25Mb, 325Mb).

Change-Id: I5fb061d781a332bd12e26807719d6cde3af34492