Allow lint's field and method lookup to include inherited members
This CL adds an "includeInherited" parameter to the getMethod and
getField methods on resolved classes. This is necessary such that
clients can deliberately decide whether they want these to be
included; prior to this change, the ECJ based lookup methods did not
include them, and the IntelliJ ones did. Furthermore, the next lint CL
requires inherited methods to be included.

Change-Id: I8610282d2f17161cf977dea80d045d905c8b5073