Fix issue with abi split when there's no universal pkg.
The issue came from the process resource task that
generates the R class. Only one of the outputs' task
is setup to generate the R class, but it only
happened if an output had no filter for abi or density.

Now we just find the first one to not have a density filter
and use it.

Also, converted ndkSanAngeles sample to use the new split
mechanism.

Change-Id: I44becc559b90b53f77561c68020725e60daa749a