Add markup format to error messages
Lint already supports marked up text; issue explanations can already
contain markup like *bold* and `symbol`, and it can convert this to
plain text or HTML markup. The latter is used not just in HTML reports
but in IDE reports.

Now that we support HTML markup in IDE tooltip messages, this CL
extends this facility from issue descriptions and explanations to the
lint error messages and location descriptions as well, and formalizes
the concept a bit more. Rather than the output format being an enum
inside the Issue class specialized as parameters to the getDescription
and getExplanation methods, it is now a top-level class itself,
TextFormat, which has methods to convert between the different formats
(raw, text, tml) and the text format is passed along to the reporting
clients.

These changes are pretty small; most of the changes in this CL simply
mark up symbols in error messages, so instead of saying
 Overriding method should call super.onDetachedFromWindow
it now says
 Overriding method should call `super.onDetachedFromWindow`
which in the IDE will be rendered as
 Overriding method should call <code>super.onDetachedFromWindow</code>

Change-Id: Ifb87c3cd1e9f0fd68854fd1e6a0c136c083da5d1