Remove the optional data attribute for lint errors
Lint has a concept of "additional data" in its reporting API.  The
intent here was to support IDE quickfixes such that they don't have to
go and dig details out of error messages. For example, if you are
reporting that the class CalendarView requires API 11, in addition to
the error message reported to the user, the lint detector would pass
"android.widget.CalendarView" and "11" as optional data. That way an
IDE quickfix doesn't have to dig this out of the error message (which
can be sensitive to the exact format of the error message).

However, this did not work out in practice. In Eclipse for example
each lint error is turned into an IMarker, which is then managed and
persisted with the IDE and there is no easy way to store additional
attributes for these. Therefore, quickfixes had to rely on strings
anyway, so none of the 200 lint rules supply this optional data at
this point. However, it's part of the API so every since report call
has an extra null parameter.

Since we're making a couple of other API changes to the reporting API
(to support markup) this seemed like the right time to also clean this
up.

Change-Id: Idc957f1170015fde7c12f709312f0fd5ba4a5151