Write test to check debug/release file name.
This is to test the fix done in c6ee0e where
I fixed the name of the debug/release library to not
be the same.

This test looks at both app and lib projects and compare
the filename of the debug and release output file to make
sure they are not the same.

Change-Id: Ie5891c73bce9804df31d3c40087f4e9b5b8db79c