Fix issue with with proguard config from aar.
This was a timing issue between configuring the proguard
task with config files coming from aars and unzipping the aars.

The configuration only included the config files from aars
if the files already existed, which is not possible on first
run when the aars haven't been unzipped yet.

Since proguard does not care about file existence during
configuration, we can include the files even if they
don't exist.

Issue 69625

Change-Id: Ic77c70c686f67863f88004d57ded867f74b81c38