Fix proguard's handling of config files. again.
My previous fix setup the proguard files with all
possible config files, not knowing which one would
exist or not. It turns out that the tests we have
for this didn't include a case where an aar does not
have a config file.

Proguard requires all files passed during setup to actually
exist during executions.

As a first fix, I'm moving the setup in the doFirst of the
task where we know that the files either exist or don't. I
also fixed our test project to better test this area.

However this uncovered an issue that using doFirst means
that that proguard task will not use changes in config
files to trigger a task execution. If only config files
are changed, the task is considered up to date. This is
a bigger problem to fix that will come in a later fix.

Change-Id: Iecafa7181b0f827eb9f78418acaf98df8e632839