Add Features to layoutlib-api
Features are designed to replace Capabilities. This uses a list of
integers so that LayoutLib can have features that the IDE doesn't know
about. This allows updating LayoutLib which supports new features
without updating the IDE. However, for taking advantage of the newly
added features, the IDE may have to be updated.

The existing capabilities are moved to features with their
ordinals as the int values.

Change-Id: I3634a7b7fd1ee59c2becd1e778043d4e2d078b47