Extract the API version from error message lookup into lint detector
For other lint detectors, the extract-information-out-of-message code
lives in the detector, not in the IDE. The API detector on the other hand
had that code on the IDE side. This CL moves it into the lint detector,
and adds unit tests which verify that all generated messages are properly
parsed.

Change-Id: I908daaf13133bcfaf14a14e3680180628242d597