Fix model for aar with local jar.
When the model is built, the aar may not have been
exploded which means the method that returns the aar
dynamically will return an empty list.

Only when building the model, we now go and manually
look inside the aar archive to find local jars if
needed (sometimes the model is built after a project
build and the aar has been exploded).

Change-Id: Ia021b52e50c86992fb5f8b35b444a517f8b00d76