Fix override detection for src/target compatibility.
When using the DSL like this:
    compileOptions.sourceCompatibility JavaVersion.VERSION_1_7
instead of
    compilOptions {
        sourceCompatibility JavaVersion.VERSION_1_7
    }

The method compileOptions(Action<>) in BaseExtension is
not called. Instead, it uses the getCompileOptions getter
and then manually configures the object with the given
closure.

This prevented setting setExplicitely flag to true, and defaulted
the source/target mode based on the API level.

Change-Id: Ie926c00b4a40d6865a317e164c956c0269fbd247